still working on config

+++ parallelDownloadThumbnails.py
@@ -1,24 +1,29 @@
 import requests
 import threading
 import os
-
-
-
 class ThumbnailDownloader:
     NUMBER_OF_THREADS = 8
     threads = []
 
     def __init__(self, jsonList):  # Input is all entries
         self.number_of_thumbnails = len(jsonList)
-
-        with open(self.getConfigFile()) as file_in:
-            lines = []
-            for line in file_in:
-                lines.append(line)
-            line = lines[0] # on the first line there should be the number.
-
-
         self.jsonList = jsonList
+        # general idea:
+        # Read the file. if it exists, read the number. if the self.numer_of_numbnails is higher than
+        # the number in file, proceed, else break if it doesn't exist break
+
+        if not os.path.exists(os.getcwd() + "config"):
+            os.mkdir("config")
+
+        else:
+            completeName = self.getfullPathConfigFile()
+
+
+
+
+
+
+
         for i in range(self.NUMBER_OF_THREADS):
             t = threading.Thread(target=self.do_request)
             t.daemon = True
@@ -47,35 +52,44 @@ class ThumbnailDownloader:
             file_extension = os.path.splitext(thumbnailURL)[1]
             ID = getID(Videourl)
             file_name = ID + file_extension
+            if not os.path.exists(os.getcwd() + "thumbnails"):
+                os.mkdir("thumbnails")
             thumbnailDirectory = os.path.join(os.getcwd(), "thumbnails")
             completeName = os.path.join(thumbnailDirectory, file_name)
-
-            headers = {
-                'user-agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/75.0.3770.142 Safari/537.36'
-            }
-
-            try:
-                r = requests.get(thumbnailURL, headers=headers)
-            except:
-                print("Requests Failed!")
-
-            try:
-                with open(completeName, 'wb') as f:
-                    f.write(r.content)
-            except:
-                print("error writing File")
-            print("It Worked :)")
+            if not os.path.exists(completeName): # if the thumbnail is already present
+                headers = {
+                    'user-agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/75.0.3770.142 Safari/537.36'
+                }
+                try:
+                    r = requests.get(thumbnailURL, headers=headers)
+                except:
+                    print("Requests Failed!")
+                try:
+                    with open(completeName, 'wb') as f:
+                        f.write(r.content)
+                except:
+                    print("error writing File")
+                print("It Worked :)")
+            else:
+                print("thumbnail exists already. No need to download. ")
+                return
 
     def writeConfigFile(self):
-
-            #  TODO: check if config file already exists.
             path = os.path.join(os.getcwd(), "config")
             completeName = os.path.join(path, "config.txt")
             with open(completeName, 'a') as f:
                 f.write(str(self.number_of_thumbnails) + '\n')
+            print("written config File")
 
-    def getConfigFile(self):
-            path = os.path.join(os.getcwd(), "config")
-            completeName = os.path.join(path, "config.txt")
+    def readConfigFile(self): # assuming that config.txt exists
+        with open(self.getfullPathConfigFile()) as file_in:
+            lines = []
+            for line in file_in:
+                lines.append(line)
+            return lines[0] # return the first Line
+
+    def getfullPathConfigFile(self):
+            configPath = os.path.join(os.getcwd(), "config")
+            completeName = os.path.join(configPath, "config.txt")
             return completeName
